utils: share the deterministic fuzzing generator
evm-fuzz's seeded generator moves to utils so that more than one fuzzer can draw
from the same stream. evm-fuzz keeps its names as aliases and is otherwise
unchanged.
